+++ db/migrate/20171109211942_create_space_ratings.rb
@@ -3,7 +3,7 @@ class CreateSpaceRatings < ActiveRecord::Migration[5.1]
     create_table :space_ratings do |t|
       t.integer :booking_id
       t.text :comment
-      t.integer :score
+      t.decimal :score
 
       t.timestamps
     end
